Split the wiji token into two contracts
The first is the token contract, the second that of the ICO.
All occurences of _totalSupply in the new sale contract were replaced by calls to the wiji contract.

The Issue event has been removed, it was redundant with OpenZeppelin's Transfer event.
Some functions have been (partly) commented out, because they need re-writing.