Implement maximum reserves in StagedPriceCrowdsale
A stage may have a reserve of tokens (or maximum ehter) that, once depleted (resp.
reached), closes the current stage and begins the next one.

If a purchase would deplete a stage and still have ether left after purchasing that,
the function will use the remaining ether to buy tokens at the price of the next
stage, recursively.

It is possible to mix tokens and ether limitations.
A reserve of 0 means unlimited tokens may be bought during the stage.

The code pertaining to the varying Wiji token rate in wiji_sale has been removed in
favour of this system.